Added experiment to test distribution effect
A new experiment was added such that a 2 digit integer value passed as
argument is taken as the percentage of managers in a set of agents. And
the effect of this will be tested. Also, modifications were made to the
other two experiments and the main engine to handle argumentfor number of
managers.